pyproject.toml
[tool.poetry]
name = "bot-cyber-detection"
version = "0.1.0"
description = ""
authors = ["kdelinx <kdelinx@gmail.com>"]
license = "None"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
sentry-sdk = {version = "^2.5.1", extras = ["httpx", "asyncpg"]}
python-telegram-bot = {version = "^21.2", extras = ["http2", "callback-data", "passport"]}
httpx = "^0.27.0"
aiofiles = "^23.2.1"
structlog = "^24.1.0"
asyncpg = "^0.29.0"
orjson = "^3.10.4"
phonenumbers = "^8.13.39"

[tool.poetry.group.dev.dependencies]
ipython = "^8.24.0"
ruff = "^0.4.5"
isort = "^5.13.2"
colorama = "^0.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79

[format]
indent-style = "tab"
quote-style = "single"

[tool.ruff.lint.isort]
length-sort = true
case-sensitive = true

[tool.ruff.lint]
# Enable Pyflakes, isort, flake8-print
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = []
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".direnv",
    ".eggs",
    ".git",
    ".mypy_cache",
    ".nox",
    ".pytype",
    ".ruff_cache",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "venv",
    "migrations",
]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

# Ignore `F401`,`F403` (import violations) in all `__init__.py` files.
# unused-import, undefined-local-with-import-star
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F403", "F401"]
